requirements.txt
# Core
snaptrade-python-sdk>=11.0.0
yfinance>=0.2.0
httpx>=0.27.0
python-dotenv>=1.0.0
loguru>=0.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
